ai-ml-work/requirements.txt
ultralytics
opencv-python
pytest
Flask
python-dotenv
requests
supabase